requirements.txt
astroid==2.3.3
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
easyocr==1.2.1
imageio==2.9.0
isort==4.3.21
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.46
Pillow==8.0.1
pylint==2.4.4
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.12.8
torch==1.7.1
torchvision==0.5.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
